+++ code/src/cljs/sixsq/nuvla/ui/main/events.cljs
@@ -304,7 +304,7 @@
                 :sixsq.nuvla.ui.profile.spec/subscription]} :db} [_ dispatch-vector]]
     (let [active? (boolean
                     (or (and (some? stripe)
-                             (#{"trialing" "active"} (:status subscription)))
+                             (#{"trialing" "active" "past_due"} (:status subscription)))
                         (nil? stripe)))]
       {:dispatch (if active?
                    dispatch-vector